Move Khronos function to cycle cog
