+++ cogs/reminders.py
@@ -21,15 +21,6 @@ class Reminders(commands.Cog, name='Reminders'):
         self.bot = bot
         self.check_reminders.start()
 
-        self.channel_time_once.start()
-        self.channel_time.start()
-
-        self.channel_phase_once.start()
-        self.channel_phase.start()
-
-        self.channel_phase_end_once.start()
-        self.channel_phase_end.start()
-
     @commands.command(
         brief='Have the bot remind you about something',
         help=(f'Usages: &remind [number] [second(s)|minute(s)|hour(s)|day(s)|week(s)] <message>\n'
@@ -112,84 +103,6 @@ class Reminders(commands.Cog, name='Reminders'):
                                                       userAt=userAt, createdAt=createdAt, message=_msg))
 
             log.info(reminders.unset_reminder(rowId, overrideId=True))
-
-    @tasks.loop(count=1)
-    async def channel_time_once(self):
-        '''Run channel_phase immediately on start before starting the actual loop'''
-        return await self.channel_time()
-
-    @tasks.loop(minutes=10)
-    async def channel_time(self):
-        '''
-        Sets a datestring as the name of a specific configured voice channel.
-        '''
-        # Get the current datetime string
-        datestring = nomic_time.get_formatted_date_string()
-
-        # Update the channel name
-        channel = await self.bot.fetch_channel(config.UTC_UPDATE_CHANNEL)
-        await channel.edit(name=datestring)
-
-    @channel_time.before_loop
-    async def before_channel_time(self):
-        '''
-        Delays the start of the time tracking loop until we get to the next 10-minute increment
-        '''
-        seconds_to_start = nomic_time.seconds_to_next_10_minute_increment()
-        log.info(f'Seconds to start tracking time: {seconds_to_start}')
-        await asyncio.sleep(seconds_to_start)
-
-    @tasks.loop(count=1)
-    async def channel_phase_once(self):
-        '''Run channel_phase immediately on start before starting the actual loop'''
-        return await self.channel_phase()
-
-    @tasks.loop(hours=24)
-    async def channel_phase(self):
-        '''
-        Sets the current phase as the name of a specific configured voice channel.
-        '''
-        # Get the current phase string
-        phasestring = nomic_time.get_current_phase_string()
-
-        # Update the channel name
-        channel = await self.bot.fetch_channel(config.PHASE_UPDATE_CHANNEL)
-        await channel.edit(name=phasestring)
-
-    @channel_phase.before_loop
-    async def before_channel_phase(self):
-        '''
-        Delays the start of the time tracking loop until the beginning of the next day
-        '''
-        seconds_to_start = nomic_time.seconds_to_next_day()
-        log.info(f'Seconds to start tracking phase: {seconds_to_start}')
-        await asyncio.sleep(seconds_to_start)
-
-    @tasks.loop(count=1)
-    async def channel_phase_end_once(self):
-        '''Run channel_phase_end immediately on start before starting the actual loop'''
-        return await self.channel_phase_end()
-
-    @tasks.loop(minutes=10)
-    async def channel_phase_end(self):
-        '''
-        Sets the time to the end of the current phase
-        '''
-        # Get the current string
-        channel_name = nomic_time.get_next_time_to_phase_end_string()
-
-        # Update the channel name
-        channel = await self.bot.fetch_channel(config.PHASE_END_UPDATE_CHANNEL)
-        await channel.edit(name=channel_name)
-
-    @channel_phase_end.before_loop
-    async def before_channel_phase_end(self):
-        '''
-        Delays the start of the time tracking loop until the beginning of the next day
-        '''
-        seconds_to_start = nomic_time.seconds_to_next_10_minute_increment()
-        log.info(f'Seconds to start tracking phase end: {seconds_to_start}')
-        await asyncio.sleep(seconds_to_start)
 
 
 async def handle_set_reminder(ctx, userId, createdAt, messageId, channelId, remindAfter, msg):
